Abstract MapProvider from PolygonMap
