+++ src/ext/Fintraffic/CustomStyle/styles.scss
@@ -19,6 +19,10 @@ body {
   margin-bottom: 1.75rem;
 }
 
+.page .page-content {
+  margin: 3.5rem 2rem;
+}
+
 // Navigation menu:
 
 .eds-contrast .eds-floating-button:focus {
